chore: 강의목록 불러오는 api경로 수정 - 타입선언은 타일 파일에 정의함 - api로직은 api폴더에 - 쿼리 사용하는 무한스크롤은 쿼리 폴더로 이동함
