+++ src/hooks/quries/useGetLecture.tsx
@@ -1,0 +1,16 @@
+import getLecturesApi from '@/api/lecture'
+import { useInfiniteQuery } from '@tanstack/react-query'
+
+//불러온 강의목록으로 무한대 호출하는 로직(쿼리사용)
+export default function useGetLectures() {
+  return useInfiniteQuery({
+    queryKey: ['get-lectures'],
+    queryFn: ({ pageParam }) => {
+      return getLecturesApi({ page: pageParam })
+    },
+    getNextPageParam: (lastPage, allPages) => {
+      return lastPage.next ? allPages.length + 1 : undefined
+    },
+    initialPageParam: 1,
+  })
+}
